refactor: inline landing page into home page
